Update goup_project_starter.cpp
Here is what I was able to come up with for the spell checker. I took the original word list, created a "library" vector, and a second vector that makes an uppercase copy of the word list. I then compare the upper case vector to the original list and have it print out all that's "Misspelled"